+++ Assets/Use Case Samples/Loot Boxes With Cooldown/Scripts/EconomyManager.cs
@@ -26,10 +26,22 @@ namespace UnityGamingServicesUseCases
                 }
             }
 
+            public async Task RefreshEconomyConfiguration()
+            {
+                // Calling GetCurrenciesAsync (or GetInventoryItemsAsync), in addition to returning the appropriate
+                // Economy configurations, will update the cached configuration list, including any new Currency, 
+                // Inventory Item, or Purchases that have been published since the last time the player's configuration
+                // was cached.
+                // 
+                // This is important to do before hitting the Economy or Remote Config services for any other calls as
+                // both use the cached data list.
+                await EconomyService.Instance.Configuration.GetCurrenciesAsync();
+            }
+
             public async Task RefreshCurrencyBalances()
             {
-                var options = new PlayerBalances.GetBalancesOptions { ItemsPerFetch = 100 };
-                var getBalancesTask = Economy.PlayerBalances.GetBalancesAsync(options);
+                var options = new GetBalancesOptions { ItemsPerFetch = 100 };
+                var getBalancesTask = EconomyService.Instance.PlayerBalances.GetBalancesAsync(options);
                 var balances = await Utils.ProcessEconomyTaskWithRetry(getBalancesTask);
 
                 // Check that scene has not been unloaded while processing async wait to prevent throw.
@@ -40,8 +52,8 @@ namespace UnityGamingServicesUseCases
 
             public async Task RefreshInventory()
             {
-                var options = new PlayerInventory.GetInventoryOptions { ItemsPerFetch = 100 };
-                var getInventoryTask = Economy.PlayerInventory.GetInventoryAsync(options);
+                var options = new GetInventoryOptions { ItemsPerFetch = 100 };
+                var getInventoryTask = EconomyService.Instance.PlayerInventory.GetInventoryAsync(options);
                 var inventory = await Utils.ProcessEconomyTaskWithRetry(getInventoryTask);
 
                 if (this == null) return;