+++ Assets/Use Case Samples/Starter Pack/Scripts/StarterPackSceneManager.cs
@@ -35,6 +35,16 @@ namespace UnityGamingServicesUseCases
 
                     Debug.Log($"Player id: {AuthenticationService.Instance.PlayerId}");
 
+                    // Economy configuration should be refreshed every time the app initializes.
+                    // Doing so updates the cached configuration data and initializes for this player any items or
+                    // currencies that were recently published.
+                    // 
+                    // It's important to do this update before making any other calls to the Economy or Remote Config
+                    // APIs as both use the cached data list. (Though it wouldn't be necessary to do if only using Remote
+                    // Config in your project and not Economy.)
+                    await EconomyManager.instance.RefreshEconomyConfiguration();
+                    if (this == null) return;
+
                     await Task.WhenAll(EconomyManager.instance.RefreshCurrencyBalances(),
                         EconomyManager.instance.RefreshInventory(),
                         RefreshStarterPackStatus());
@@ -84,7 +94,7 @@ namespace UnityGamingServicesUseCases
                 { 
                     StarterPackSampleView.instance.SetInteractable(false);
 
-                    var balanceResponse = await Economy.PlayerBalances.IncrementBalanceAsync("GEM", 10);
+                    var balanceResponse = await EconomyService.Instance.PlayerBalances.IncrementBalanceAsync("GEM", 10);
                     if (this == null) return;
 
                     EconomyManager.instance.SetCurrencyBalance(balanceResponse.CurrencyId, balanceResponse.Balance);