+++ Assets/Use Case Samples/Virtual Shop/Scripts/EconomyManager.cs
@@ -22,6 +22,11 @@ namespace UnityGamingServicesUseCases
             public Dictionary<string, (List<ItemAndAmountSpec> costs, List<ItemAndAmountSpec> rewards)>
                 virtualPurchaseTransactions { get; private set; }
 
+            public List<CurrencyDefinition> currencyDefinitions { get; private set; }
+            public List<InventoryItemDefinition> inventoryItemDefinitions { get; private set; }
+
+            List<VirtualPurchaseDefinition> m_VirtualPurchaseDefinitions;
+
             void Awake()
             {
                 if (instance != null && instance != this)
@@ -42,15 +47,37 @@ namespace UnityGamingServicesUseCases
                 }
             }
 
+            public async Task RefreshEconomyConfiguration()
+            {
+                // Calling GetCurrenciesAsync (or GetInventoryItemsAsync, or GetVirtualPurchasesAsync, etc), in addition
+                // to returning the appropriate Economy configurations, will update the cached configuration list,
+                // including any new Currency, Inventory Item, or Purchases that have been published since the last
+                // time the player's configuration was cached.
+                // 
+                // This is important to do before hitting the Economy or Remote Config services for any other calls as
+                // both use the cached data list.
+                var getCurrenciesTask = EconomyService.Instance.Configuration.GetCurrenciesAsync();
+                var getInventoryItemsTask = EconomyService.Instance.Configuration.GetInventoryItemsAsync();
+                var getVirtualPurchasesTask = EconomyService.Instance.Configuration.GetVirtualPurchasesAsync();
+
+                await Task.WhenAll(getCurrenciesTask, getInventoryItemsTask, getVirtualPurchasesTask);
+
+                // Check that scene has not been unloaded while processing async wait to prevent throw.
+                if (this == null) return;
+
+                currencyDefinitions = getCurrenciesTask.Result;
+                inventoryItemDefinitions = getInventoryItemsTask.Result;
+                m_VirtualPurchaseDefinitions = getVirtualPurchasesTask.Result;
+            }
+
             public async Task RefreshCurrencyBalances()
             {
                 try
                 {
-                    var options = new PlayerBalances.GetBalancesOptions { ItemsPerFetch = 100 };
-                    var getBalancesTask = Economy.PlayerBalances.GetBalancesAsync(options);
+                    var options = new GetBalancesOptions { ItemsPerFetch = 100 };
+                    var getBalancesTask = EconomyService.Instance.PlayerBalances.GetBalancesAsync(options);
                     var balances = await Utils.ProcessEconomyTaskWithRetry(getBalancesTask);
 
-                    // Check that scene has not been unloaded while processing async wait to prevent throw.
                     if (this == null) return;
 
                     currencyHudView.SetBalances(balances);
@@ -69,8 +96,8 @@ namespace UnityGamingServicesUseCases
 
                 try
                 {
-                    var options = new PlayerInventory.GetInventoryOptions { ItemsPerFetch = 100 };
-                    var getInventoryTask = Economy.PlayerInventory.GetInventoryAsync(options);
+                    var options = new GetInventoryOptions { ItemsPerFetch = 100 };
+                    var getInventoryTask = EconomyService.Instance.PlayerInventory.GetInventoryAsync(options);
                     var getInventoryResult = await Utils.ProcessEconomyTaskWithRetry(getInventoryTask);
 
                     if (this == null) return;
@@ -84,15 +111,17 @@ namespace UnityGamingServicesUseCases
                 }
             }
 
-            public async Task ReadAllVirtualPurchaseTransactions()
+            public void InitializeVirtualPurchaseLookup()
             {
-                var virtualPurchaseDefinitions = await Economy.Configuration.GetVirtualPurchasesAsync();
-                if (this == null) return;
+                if (m_VirtualPurchaseDefinitions == null)
+                {
+                    return;
+                }
 
                 virtualPurchaseTransactions = new Dictionary<string, 
                     (List<ItemAndAmountSpec> costs, List<ItemAndAmountSpec> rewards)>();
 
-                foreach (var virtualPurchaseDefinition in virtualPurchaseDefinitions)
+                foreach (var virtualPurchaseDefinition in m_VirtualPurchaseDefinitions)
                 {
                     var costs = ParseEconomyItems(virtualPurchaseDefinition.Costs);
                     var rewards = ParseEconomyItems(virtualPurchaseDefinition.Rewards);
@@ -118,7 +147,7 @@ namespace UnityGamingServicesUseCases
             {
                 try
                 {
-                    return await Economy.Purchases.MakeVirtualPurchaseAsync(virtualPurchaseId);
+                    return await EconomyService.Instance.Purchases.MakeVirtualPurchaseAsync(virtualPurchaseId);
                 }
                 catch (EconomyException e)
                 when (e.ErrorCode == k_EconomyPurchaseCostsNotMetStatusCode)
@@ -139,7 +168,7 @@ namespace UnityGamingServicesUseCases
             {
                 try
                 {
-                    await Economy.PlayerBalances.IncrementBalanceAsync(currencyId, amount);
+                    await EconomyService.Instance.PlayerBalances.IncrementBalanceAsync(currencyId, amount);
                 }
                 catch (Exception e)
                 {